+++ frontend/src/pages/index.ts
@@ -1,7 +1,6 @@
 // Export all pages
 export { default as BacktestPage } from './BacktestPage';
 export { default as BBMidPage } from './BBMidPage';
-export { default as ComboFilterPage } from './ComboFilterPage';
 export { default as TrendJudgmentPage } from './TrendJudgmentPage';
 export { default as TrendChartPage } from './TrendChartPage';
 export { default as TradeAnalysisPage } from './TradeAnalysisPage';